Allow non-sequential directory offsets
The FUSE protocol allows each directory entry to include a off field in
struct fuse_dirent. It's purpose is to uniquely identify a directory
entry so that a future FUSE_READDIR operation may continue where a
previous one left off. The FUSE file system is allowed to choose the
format of those off values. In particular, sequentially increasing
indexes are usually not adequate, because if a directory entry is
deleted in between two FUSE_READDIR operations, using sequential off
values could lead to skipping over a directory entry.

Instead of attempting to manage those offsets with fuse3, forcing them
to be sequential, pass responsibility to the file system.

Fixes #12